requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.215.1
databases[postgresql]==0.8.0
asyncpg==0.29.0
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
